Remove appData module and migration routine
