+++ src/main.js
@@ -1,8 +1,10 @@
+import path from 'path';
+
 import { app } from 'electron';
 import setupElectronReload from 'electron-reload';
+import jetpack from 'fs-jetpack';
 
 import { setupErrorHandling } from './errorHandling';
-import appData from './main/appData';
 import './main/basicAuth';
 import { processDeepLink } from './main/deepLinks';
 import { getMainWindow } from './main/mainWindow';
@@ -25,13 +27,23 @@ async function prepareApp() {
 	app.setAsDefaultProtocolClient('rocketchat');
 	app.setAppUserModelId('chat.rocket');
 
-	await appData.initialize();
+	const dirName = process.env.NODE_ENV === 'production' ? app.name : `${ app.name } (${ process.env.NODE_ENV })`;
+
+	app.setPath('userData', path.join(app.getPath('appData'), dirName));
+
+	if (process.argv[2] === '--reset-app-data') {
+		const dataDir = app.getPath('userData');
+		await jetpack.removeAsync(dataDir);
+		app.relaunch({ args: [process.argv[1]] });
+		app.exit();
+		return;
+	}
 
 	const canStart = process.mas || app.requestSingleInstanceLock();
 
 	if (!canStart) {
-		app.quit();
-		return false;
+		app.exit();
+		return;
 	}
 
 	app.commandLine.appendSwitch('--autoplay-policy', 'no-user-gesture-required');
@@ -53,17 +65,15 @@ async function prepareApp() {
 	app.on('second-instance', (event, argv) => {
 		argv.slice(2).forEach(processDeepLink);
 	});
-
-	return true;
 }
 
 (async () => {
-	if (!await prepareApp()) {
-		return;
-	}
+	await prepareApp();
 
 	await app.whenReady();
+
 	await setupI18next();
+
 	app.emit('start');
 	await getMainWindow();
 	process.argv.slice(2).forEach(processDeepLink);
